# Improve Regex performance (mainly interpreted)