--- a/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexCharClass.cs
+++ b/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexCharClass.cs
@@ -732,7 +732,7 @@ namespace System.Text.RegularExpressions
         /// </summary>
         public static char SingletonChar(string set)
         {
-            Debug.Assert(IsSingleton(set) || IsSingletonInverse(set), "Tried to get the singleton char out of a non singleton character class");
+            Debug.Assert(IsSingletonInverse(set), "Tried to get the singleton char out of a non singleton character class");
             return set[SetStartIndex];
         }
 
@@ -746,14 +746,6 @@ namespace System.Text.RegularExpressions
             charClass[SetLengthIndex] == 0 &&
             !IsNegated(charClass) &&
             !IsSubtraction(charClass);
-
-        /// <summary><c>true</c> if the set contains a single character only</summary>
-        public static bool IsSingleton(string set) =>
-            set[CategoryLengthIndex] == 0 &&
-            set[SetLengthIndex] == 2 &&
-            !IsNegated(set) &&
-            !IsSubtraction(set) &&
-            (set[SetStartIndex] == LastChar || set[SetStartIndex] + 1 == set[SetStartIndex + 1]);
 
         public static bool IsSingletonInverse(string set) =>
             set[CategoryLengthIndex] == 0 &&