--- a/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexNode.cs
+++ b/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexNode.cs
@@ -279,8 +279,7 @@ namespace System.Text.RegularExpressions
         }
 
         /// <summary>
-        /// Simple optimization. If a set is a singleton, an inverse singleton,
-        /// or empty, it's transformed accordingly.
+        /// Simple optimization. If a set is an inverse singleton or empty, it's transformed accordingly.
         /// </summary>
         private RegexNode ReduceSet()
         {
@@ -292,12 +291,6 @@ namespace System.Text.RegularExpressions
             {
                 NType = Nothing;
                 Str = null;
-            }
-            else if (RegexCharClass.IsSingleton(Str))
-            {
-                Ch = RegexCharClass.SingletonChar(Str);
-                Str = null;
-                NType += (One - Set);
             }
             else if (RegexCharClass.IsSingletonInverse(Str))
             {