… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexRunner.cs