@danmosemsft 2019-12-03
This is awesome, thanks @stephentoub. Do you think there may much more that can be achieved incrementally?