@stephentoub 2019-12-03
> Do you think there may much more that can be achieved incrementally?

I do, e.g. https://github.com/dotnet/runtime/pull/496